% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLigandsReceptors.R
\name{getLigandsReceptors}
\alias{getLigandsReceptors}
\title{Retrieve a database of ligand-receptor pairs}
\usage{
getLigandsReceptors(version = "latest",
  cellularCompartment = c("Membrane", "ECM", "Both", "Secreted"),
  manualAnnotation = "Correct", ligandClass = c("Other", "Cytokine",
  "Chemokine", "GrowthFactor", "Interleukin"))
}
\arguments{
\item{version}{Currently supports the following values: \itemize{
  \item{latest} points to \code{2.1.0}
  \item{1.0.0} contains manual annotation for all genes expressed in bone marrow. This version was used for analysis is the Baccin et al paper. Information on heterodimeric receptors is only included for integrins.
  \item{2.0.0} contains manual annotation for all genes in the geneome (not yet complete).
  \item{2.1.0} integrates heterodimer information for non-integrin receptors from cellphoneDB.
  \item{3.0.0} adds a translation of the complete cellphoneDB to mouse homologues (not tested).
  \item Alternatively, a data frame with the same column names as \code{ligandsReceptors_1.0.0} can be used.
}}

\item{cellularCompartment}{A character vector to select permitted localizations of the \strong{ligand}. Valid entries are \code{Membrane}, \code{ECM}, \code{Secreted} and \code{Both}, which refers to membrane-bound ligands that can also be secreted.}

\item{manualAnnotation}{A character vector to select permitted annotation status of the interaction. Valid entries are \code{Correct}, \code{Incorrect}, \code{Not Expressed} (i.e. entries not annotated in v1.0.0) and \code{Irrelevant} (i.e. entries we consider correct but not relevant at homeostasis, e.g. interactions involving activated components of the complement system)}

\item{ligandClass}{A character vector to select permitted classes of ligands. Defaults to all, i.e. \code{c("Other","Cytokine","Chemokine","GrowthFactor","Interleukin")}}
}
\value{
Returns a data frame with 11 columns: \itemize{
\item{Pair.Name}{Unique identifier for the ligand-receptor pair}
\item{Ligand.Mouse}{MGI symbol of the ligand. }
\item{Receptor.Mouse}{MGI symbol of the receptor. In case multiple receptors are required for complex formation, multiple MGI symbols separated by \strong{&}, as in Itgal&Itgb2. }
\item{Source}{Ramilowski for entries taken from \emph{Ramilowski et al., Nature Communications 2015}; Baccin for entries added as part of this work}
\item{Reference}{Pubmed ID or similar}
\item{ManualAnnotation}{The literature evidence for all entries from Ramilowski et al with expression in bone where checked manually.}
\item{Ligand.CC}{Cellular Compartment of the ligand - one of Secreted, Membrane, Both, or ECM. Annotation is based on Gene Ontology, but was checked manually for entries with expression in bone}
\item{Ligand.GO}{Classification of the ligand by gene ontology into one of the following classes: GrowthFactor, Cytokine, Chemokine, Interleukin, or Other}
}
}
\description{
Retrieve a database of ligand-receptor pairs
}
