% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLigandsReceptors.R
\name{getRNAMagnetGenes}
\alias{getRNAMagnetGenes}
\title{Extracts receptor-ligand pairs mediating interaction between two cell types}
\usage{
getRNAMagnetGenes(rnamagnet, pop_l, pop_r, thresh = 0.05)
}
\arguments{
\item{rnamagnet}{An object of class \code{\link{rnamagnet}}}

\item{pop_l}{Sending population to consider}

\item{pop_r}{Receiving population to consider}

\item{thresh}{Threshold score for ligand and receptor expression}
}
\value{
A data frame conaining the score and the receptor-ligand pair
}
\description{
Extracts receptor-ligand pairs mediating interaction between two cell types
}
