% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NicheData10x}
\alias{NicheData10x}
\title{10x dataset from mouse bone marrow}
\format{An object of class \code{\link[Seurat]{seurat}}, containing clustering and dimensionality reduction}
\usage{
NicheData10x
}
\description{
10x dataset from mouse bone marrow
}
\keyword{datasets}
