% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NicheDataLCM}
\alias{NicheDataLCM}
\title{LCM dataset from mouse bone marrow}
\format{Read count matrix, see \code{\link{NicheMetaDataLCM}} for per-sample metadata.}
\usage{
NicheDataLCM
}
\description{
LCM dataset from mouse bone marrow
}
\keyword{datasets}
