% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NicheMarkers10x}
\alias{NicheMarkers10x}
\title{Markers for populations defined by 10x genomics}
\format{Data frame of cell type specific markers - the output of running \code{FindMarkersAll{\link{NicheData10x}, method="roc"}}}
\usage{
NicheMarkers10x
}
\description{
Markers for populations defined by 10x genomics
}
\keyword{datasets}
