% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunctions.R
\name{PlotSignalingNetwork}
\alias{PlotSignalingNetwork}
\title{Visualizes RNAMagnet output as a signaling network}
\usage{
PlotSignalingNetwork(rnamagnet, threshold = 0.01, colors = NULL,
  useLabels = T)
}
\arguments{
\item{rnamagnet}{An object of class \code{\link{rnamagnet}}, as produced by \code{\link{RNAMagnetSignaling}}}

\item{threshold}{Threshold interaction strength for including an edge in the network.}

\item{colors}{A named vector of colors to be used for plotting; names correspond to levels of \code{rnamagnet@class}}

\item{useLabels}{Plot the population labels on the graph?}

\item{return}{Determines object to return; one of "summary" or "rnamagnet-class"}
}
\value{
Plots the network and invisibly returns an object of class \code{\link[igraph]{igraph}}
}
\description{
Visualizes RNAMagnet output as a signaling network
}
