% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnamagnet-class.R
\docType{class}
\name{rnamagnet-class}
\alias{rnamagnet-class}
\alias{rnamagnet}
\title{RNAMagnet class}
\description{
This class contains representation for the output of \code{\link{RNAMagnetBase}}.
}
\section{Slots}{

\describe{
\item{\code{similarity}}{Cell-cell distance matrix}

\item{\code{anchors}}{Mean gene expression matrix of anchor populations}

\item{\code{interaction}}{Interaction score matrix between cells and anchor populations}

\item{\code{specificity}}{Specificity of interaction scores relative to similar cells}

\item{\code{adhesiveness}}{Total interaction potential per cell}

\item{\code{celltype}}{Per-cell cell type annotation, copied over from \code{seurat@ident}}

\item{\code{mylr}}{The output of \code{\link{getLigandsReceptors}} that this instance of RNAMagnet is based on}

\item{\code{params}}{Named list of paramers used in \code{\link{RNAMagnetBase}}}
}}

