% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAMagnet.R
\name{RNAMagnetAnchors}
\alias{RNAMagnetAnchors}
\title{Runs RNAMagnet for identifying localization of single cells to anchor populations}
\usage{
RNAMagnetAnchors(seurat, anchors, return = "summary",
  neighborhood.distance = 0.7, neighborhood.gradient = 3, .k = 10,
  .x0 = 0.5, .minExpression = 0, .version = "latest",
  .cellularCompartment = c("Membrane", "ECM", "Both"),
  .manualAnnotation = "Correct")
}
\arguments{
\item{seurat}{An object of class \code{\link[Seurat]{seurat}} containing a valid clustering and t-SNE information. For information on how to create such an object, see https://satijalab.org/seurat/get_started.html}

\item{anchors}{A character vector of anchor populations. Entries must be levels of \code{seurat@ident}}

\item{return}{Determines object to return; one of "summary" or "rnamagnet-class"}

\item{neighborhood.distance}{See detail}

\item{neighborhood.gradient}{See detail}

\item{...}{For explanation of all further parameters, see \code{\link{RNAMagnetBase}}.}
}
\value{
Returns a data frame containing, for each cell, the propensity to physically interact with the various anchor populations as well as an overall adhesiveness score and prefered interaction partner. Alternatively, if \code{return} ist set to \code{rnamagnet-class}, an object of class \code{\link{rnamagnet}}.
}
\description{
RNAMagnet comes in two flavors: \code{RNAMagnetAnchors} and \code{\link{RNAMagnetSignaling}}. This function is meant to identify, for each single cell from a \code{\link[Seurat]{seurat}} object, the propensity to physically adhere to a set of anchor populations. For example, this function can identify if a single cell is more likely to bind to arteriolar or sinusoidal vessels.
}
\details{
Highly similar cell types can localize to different physical structures. For example, one type of pericyte may localize to sinusoids and another type may localize to arterioles. To increase the resolution in this scenario, \code{RNAMagnetAnchors} for each pair of single cells and anchor populations therefore computes a specificity score that describes how the single cell differs from \emph{similar} cells.

In our hands, the behavior of RNAMagnet is largely insensitive to the parameters \code{neighborhood.distance} and \code{neighborhood.gradient} that define what actually constitutes a similar cell. To explore how these parameters affect the weight each single cell carries in specificity score computation, see the example code.
}
\examples{
\dontrun{
result <- RNAMagnetAnchors(NicheData10x, anchors = c("Sinusoidal ECs","Arteriolar ECs","Smooth muscle","Osteoblasts"))
qplot(x =NicheData10x@dr$tsne@cell.embeddings[,1], y=NicheData10x@dr$tsne@cell.embeddings[,2], \\
     color = direction,size=I(0.75),alpha= adhesiveness,data=result) + \\
     scale_color_brewer(name = "RNAMagnet\\nLocation",palette= "Set1") + \\
     scale_alpha_continuous(name = "RNAMagnet\\nAdhesiveness")

#To understand the effect of the neighborhood.distance and neighborhood.gradient parameters
#consider the following snippet:
myMagnet <- RNAMagnetAnchors(NicheData10x, return = "rnamagnet-class", \\
           anchors = c("Sinusoidal ECs","Arteriolar ECs","Smooth muscle","Osteoblasts"))
use <- 1234 #select some cell of interest
kernel <- function(x,k=10, x0=0.5) 1/(1+exp(-k * (x-x0))) #defines weighing function
plf <- data.frame(x = seurat@dr$tsne@cell.embeddings[,1],y = seurat@dr$tsne@cell.embeddings[,2], \\
                 weight = 1-kernel(dbig[use,],k=neighborhood.gradient,x0=neighborhood.distance))
qplot(x = x, y= y, color = scores, data=plf) + \\
    scale_color_gradientn(name = "Weight in local neighborhood", colours = c("#EEEEEE","#999999","blue","red")) + \\
    geom_point(color = "black", shape = 17, size= 3, data=plf[use,])
}
}
