% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAMagnet.R
\name{RNAMagnetBase}
\alias{RNAMagnetBase}
\title{Low-level function to run core RNA magnet steps}
\usage{
RNAMagnetBase(seurat, anchors = NULL, neighborhood.distance = NULL,
  neighborhood.gradient = NULL, .k = 10, .x0 = 0.5, .minExpression,
  .version = "latest", .cellularCompartment,
  .manualAnnotation = "Correct", .symmetric = F)
}
\arguments{
\item{seurat}{An object of class \code{\link[Seurat]{seurat}} containing a valid clustering and t-SNE information. For information on how to create such an object, see https://satijalab.org/seurat/get_started.html}

\item{anchors}{A character vector of anchor populations. Entries must be levels of \code{seurat@ident}. If \code{NULL}: All entries of  \code{seurat@ident} are used as anchors.}

\item{neighborhood.distance}{See \code{\link{RNAMagnetAnchors}}}

\item{neighborhood.gradient}{See \code{\link{RNAMagnetAnchors}}}

\item{.k}{Fuzzification parameter, see detail. Recommended to leave at the default value.}

\item{.x0}{Fuzzification parameter, see detail. Recommended to leave at the default value.}

\item{.minExpression}{Minimal expression level of genes to be included, specified as the number of cells in the dataset that express the gene.}

\item{.version}{The version of the underlying ligand-receptor database. Recommended to leave at the default value. See \code{\link{getLigandsReceptors}}.}

\item{.cellularCompartment}{Types of ligands to be included. For physical interactions, defaults to \code{ c("Membrane","ECM","Both")}.  See \code{\link{getLigandsReceptors}}.}

\item{.manualAnnotation}{Annotation status of ligands to be included. Default to \code{"Correct"}. See \code{\link{getLigandsReceptors}}.}

\item{.symmetric}{Assume that if A is a receptor for B, B is also a receptor for A}
}
\value{
Returns an object of class \code{\link{rnamagnet}}
}
\description{
Users are advised to use the top-level functions \code{\link{RNAMagnetAnchors}} and \code{\link{RNAMagnetSignaling}} which appropriately set default parameters and return user-friendly return values.
This is a low level function for development purposes.
}
\details{
The algorithm takes the following steps: \enumerate{
\item Ligand-receptor pairs are selected based on the parameters \code{.version}, \code{.cellularCompartment} and \code{.manualAnnotation}. Choice of \code{.cellularCompartment} is crucial for determining the algorithm's behavior, e.g. if set to \code{c("Secreted","Both")}, paracrine signaling interactions involving soluble ligands are investigated.
\item Dropout values in the expression levels of ligands and receptors are imputed using \code{\link[Rmagic]{magic}}
\item Mean expression level of ligands and receptors is computed for all anchor populations
\item For each cell or anchor population, the expression of each ligand and receptor is encoded as a fuzzy logic variable
\item Fuzzy logic AND is used to compute probabilities for a given interaction to be active between a single cell and an anchor population
\item An interaction score is computed as the sum of interaction probabilities across all possible ligand-receptor pairs
\item Specificty scores are computed by comparing interaction scores to average interaction scores in a local neighborhood.
}

Add the methods section of the paper here!
}
