% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NicheMarkers10x}
\alias{NicheMarkers10x}
\title{Markers for populations defined by 10x genomics}
\format{An object of class \code{data.frame} with 10330 rows and 9 columns.}
\usage{
NicheMarkers10x
}
\description{
Markers for populations defined by 10x genomics
}
\keyword{datasets}
