% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NicheMetaDataLCM}
\alias{NicheMetaDataLCM}
\title{Meta data for LCM dataset}
\format{An object of class \code{data.frame} with 76 rows and 10 columns.}
\usage{
NicheMetaDataLCM
}
\description{
Meta data for LCM dataset
}
\keyword{datasets}
